Switch to PUT to be more semantically correct
Technically this call should be idempotent as it is telling a web service that an event has happened so a PUT is more appropriate.
However this does add to a bit of extra growth in code as urllib2 will only normally GET/POST.
The alternative would be to use requests but this adds a dependency to the project which adds more complication in building and deploying a package for AWS Lambda.

Method overriding for urllib2 was found at http://stackoverflow.com/questions/21243834/doing-put-using-python-urllib2/31839324#31839324